Extract card model
