+++ kcards/models/__init__.py
@@ -1,1 +1,2 @@
-from .room import Room, Card
+from .room import Room
+from .card import Card
