+++ kcards/models/card.py
@@ -1,0 +1,13 @@
+from collections import OrderedDict
+
+
+class Card(OrderedDict):
+    """Represents a colored card raised by a person."""
+
+    def __init__(self, name, color):
+        super(Card, self).__init__()
+        self['name'] = name
+        self['color'] = color
+
+    def __repr__(self):
+        return dict.__repr__(self)
